Fix 5701: Email widget displays valid emails then lots of empty entries with 01/01/1970
In the next situation:
 - POP3 Account
 - No remove messages from mail server Email account setting
 - Single mailbox is selected

When a mail is deleted from local inbox, the mail is dropped to trash folder, but a null message
flagged with deleted status is created to allow sincronization with the pop3 server to not download
the message again. When the message is purged from server, the null message is also removed from inbox

The mail activity make use of Message.FLAG_LOADED_SELECTION when query the mail database, and is not
showing this null messages, but the email widget, on "Simple mailbox selection" is
not using Message.FLAG_LOADED_SELECTION flag to discriminate this "null messages".

Change-Id: I8886ab8efc709a6e3a378c94b3c32063eea2d518